requirements.txt
pandas
numpy
yfinance
alpaca-trade-api
python-dotenv